refactor: separate version response to dto
