+++ backend/src/main/kotlin/com/aropra/dto/VersionResponse.kt
@@ -1,0 +1,19 @@
+package com.aropra.dto
+
+import com.aropra.domain.Version
+import java.time.format.DateTimeFormatter
+
+data class VersionResponse(
+    val id: Long,
+    val version: String,
+    val updatedAt: String,
+    val description: List<String>,
+)
+
+fun Version.toResponse(): VersionResponse =
+    VersionResponse(
+        id = this.id,
+        version = this.version,
+        updatedAt = this.updatedAt.format(DateTimeFormatter.ISO_LOCAL_DATE_TIME),
+        description = this.description,
+    )
