+++ backend/src/main/kotlin/com/aropra/service/VersionService.kt
@@ -1,9 +1,9 @@
 package com.aropra.service
 
-import com.aropra.domain.Version
+import com.aropra.dto.VersionResponse
+import com.aropra.dto.toResponse
 import com.aropra.repository.VersionRepository
 import org.springframework.stereotype.Service
-import java.time.format.DateTimeFormatter
 
 @Service
 class VersionService(
@@ -13,19 +13,4 @@ class VersionService(
         versionRepository.findAllByOrderByUpdatedAtDesc().map {
             it.toResponse()
         }
-
-    private fun Version.toResponse(): VersionResponse =
-        VersionResponse(
-            id = this.id,
-            version = this.version,
-            updatedAt = this.updatedAt.format(DateTimeFormatter.ISO_LOCAL_DATE_TIME),
-            description = this.description,
-        )
 }
-
-data class VersionResponse(
-    val id: Long,
-    val version: String,
-    val updatedAt: String,
-    val description: List<String>,
-)
